Allow user to disable LOG_FILE handler
As we already set the LOG_FILE configuration variable by default, there was no
way for the user to disable having a log file. Even if she tried to do

    LOG_FILE = None

It won't work, because logging.handlers.RotatingFileHandler() throws an error
when called with None. Even if it worked, we still added this handler a few
lines down, because we were checking if there was a LOG_FILE key in the config
file (which is always true, given that we set it by default and there's no way
IMHO to delete it)

With this patch, we only create the file handler if there is a
RotatingFileHandler, and it's trueish.