Add tests for the example app
Some of these tests (and some of the core tests) are currently failing
because the tests interfere with each other because there is state that
isn't reset between tests.

Test may pass when run on their own or when just one test module is run,
but fail when all test modules are run together.